🚚 move addToManyQueues from methods to process-queue
