+++ modules/cf-core/src/methods/index.ts
@@ -1,2 +1,1 @@
 export { createRpcRouter, eventNameToImplementation, methodNameToImplementation } from "./api";
-export { addToManyQueues } from "./queued-execution";
